Gather usages of encoded filename and put it in a common function
The workspace/symbol request actually did this wrong.
